Update server.py
New calculation of points Before: Just absolute value. Now, based on the real saving reached with the purchase.  Then, as the user is "more ecological" if the savings are higher, then the relationship with the points should be direct instead of indirect. Why are the points multiplied by 100? Just to get "nicer" number. 

Reference = 0.39 Kg CO2 / 100g Basket per person. 

For the calculation, I considered the weekly average consumption in Germany per capita  "basic" products (according to some EU and Bundesregierung documents in addition to what is available in the Buhler list)

Achtung! Now a player can make negative points!